mat_po_wmod

+++ Math/matrix_pow.cpp
@@ -47,9 +47,9 @@ using mat = vector<vec>;
 
 mat mul(mat &A, mat &B) {
     mat C(A.size(),vec(B[0].size()));
-    for(int i=0;i<A.size();i++){
-        for(int k=0;k<B.size();k++){
-            for(int j=0;j<B[0].size();j++){
+    for(int i=0;i<(int)A.size();i++){
+        for(int k=0;k<(int)B.size();k++){
+            for(int j=0;j<(int)B[0].size();j++){
                 C[i][j] = C[i][j] + A[i][k]*B[k][j];
             }
         }
@@ -59,7 +59,7 @@ mat mul(mat &A, mat &B) {
 
 mat matpow(mat A,ll n){
     mat B(A.size(), vec(A.size()));
-    for(int i=0;i<A.size();i++) B[i][i] = 1;
+    for(int i=0;i<(int)A.size();i++) B[i][i] = 1;
     while(n > 0){
         if(n & 1) B = mul(B,A);
         A = mul(A,A);
